Extract pydantic-v2 JSON-Schema tools into separate module
